Check in work on nslite with RPC.
This is a work in progress. It's a bigger commit than I like but
it's time to get it into git.

The idea is to have nslite talk a variation on JSON-RPC. The packet
length is appended to the front in 2 bytes binary to make blocking
stdio read/write quite easy. It should make erlang integration really
trivial.

Once I get the rpc part working, I'll implement perl and erlang drivers.